Moved type and existence checkers to subdir SA (Syntax Analysis)
