+++ official/src/py/SA/ExistenceChecker.py
